Fix duplicating task index
